update with current branch

+++ bcbio/utils.py
@@ -1,5 +1,6 @@
 """Helpful utilities for building analysis pipelines.
 """
+import glob
 import gzip
 import os
 import tempfile
@@ -9,15 +10,18 @@ import contextlib
 import itertools
 import functools
 import random
-import ConfigParser
-import collections
 import fnmatch
 import subprocess
 import sys
 import types
 
+import six
 import toolz as tz
 import yaml
+
+from collections import Mapping, OrderedDict
+
+
 try:
     from concurrent import futures
 except ImportError:
@@ -48,7 +52,7 @@ def map_wrap(f):
     """
     @functools.wraps(f)
     def wrapper(*args, **kwargs):
-        return apply(f, *args, **kwargs)
+        return f(*args, **kwargs)
     return wrapper
 
 def transform_to(ext):
@@ -194,7 +198,18 @@ def chdir(new_dir):
 
     http://lucentbeing.com/blog/context-managers-and-the-with-statement-in-python/
     """
-    cur_dir = os.getcwd()
+    # On busy filesystems can have issues accessing main directory. Allow retries
+    num_tries = 0
+    max_tries = 5
+    cur_dir = None
+    while cur_dir is None:
+        try:
+            cur_dir = os.getcwd()
+        except OSError:
+            if num_tries > max_tries:
+                raise
+            num_tries += 1
+            time.sleep(2)
     safe_makedir(new_dir)
     os.chdir(new_dir)
     try:
@@ -221,6 +236,17 @@ def file_exists(fname):
         return fname and os.path.exists(fname) and os.path.getsize(fname) > 0
     except OSError:
         return False
+
+
+def get_size(path):
+    """ Returns the size in bytes if `path` is a file,
+        or the size of all files in `path` if it's a directory.
+        Analogous to `du -s`.
+    """
+    if os.path.isfile(path):
+        return os.path.getsize(path)
+    return sum(get_size(os.path.join(path, f)) for f in os.listdir(path))
+
 
 def file_uptodate(fname, cmp_fname):
     """Check if a file exists, is non-empty and is more recent than cmp_fname.
@@ -254,7 +280,7 @@ def read_galaxy_amqp_config(galaxy_config, base_dir):
     """Read connection information on the RabbitMQ server from Galaxy config.
     """
     galaxy_config = add_full_path(galaxy_config, base_dir)
-    config = ConfigParser.ConfigParser()
+    config = six.moves.configparser.ConfigParser()
     config.read(galaxy_config)
     amqp_config = {}
     for option in config.options("galaxy_amqp"):
@@ -313,7 +339,7 @@ def remove_plus(orig):
     """
     for ext in ["", ".idx", ".gbi", ".tbi", ".bai"]:
         if os.path.exists(orig + ext):
-             remove_safe(orig + ext)
+            remove_safe(orig + ext)
 
 def copy_plus(orig, new):
     """Copy a fils, including biological index files.
@@ -325,12 +351,15 @@ def copy_plus(orig, new):
 def symlink_plus(orig, new):
     """Create relative symlinks and handle associated biological index files.
     """
-    for ext in ["", ".idx", ".gbi", ".tbi", ".bai"]:
+    orig = os.path.abspath(orig)
+    if not os.path.exists(orig):
+        raise RuntimeError("File not found: %s" % orig)
+    for ext in ["", ".idx", ".gbi", ".tbi", ".bai", ".fai"]:
         if os.path.exists(orig + ext) and (not os.path.lexists(new + ext) or not os.path.exists(new + ext)):
             with chdir(os.path.dirname(new)):
                 remove_safe(new + ext)
-               # Work around symlink issues on some filesystems. Randomly
-               # fail to symlink.
+                # Work around symlink issues on some filesystems. Randomly
+                # fail to symlink.
                 try:
                     os.symlink(os.path.relpath(orig + ext), os.path.basename(new + ext))
                 except OSError:
@@ -339,13 +368,28 @@ def symlink_plus(orig, new):
                         shutil.copyfile(orig + ext, new + ext)
     orig_noext = splitext_plus(orig)[0]
     new_noext = splitext_plus(new)[0]
-    for sub_ext in [".bai"]:
+    for sub_ext in [".bai", ".dict"]:
         if os.path.exists(orig_noext + sub_ext) and not os.path.lexists(new_noext + sub_ext):
             with chdir(os.path.dirname(new_noext)):
                 os.symlink(os.path.relpath(orig_noext + sub_ext), os.path.basename(new_noext + sub_ext))
 
-def open_gzipsafe(f):
-    return gzip.open(f) if f.endswith(".gz") else open(f)
+def open_gzipsafe(f, is_gz=False):
+    if f.endswith(".gz") or is_gz:
+        if six.PY3:
+            return gzip.open(f, "rt", encoding="utf-8", errors="ignore")
+        else:
+            return gzip.open(f)
+    else:
+        if six.PY3:
+            return open(f, encoding="utf-8", errors="ignore")
+        else:
+            return open(f)
+
+def is_empty_gzipsafe(f):
+    h = open_gzipsafe(f)
+    is_empty = len(h.read(1)) > 0
+    h.close()
+    return is_empty
 
 def append_stem(to_transform, word):
     """
@@ -408,7 +452,7 @@ def robust_partition_all(n, iterable):
         x = []
         for _ in range(n):
             try:
-                x.append(it.next())
+                x.append(next(it))
             except StopIteration:
                 yield x
                 # Omitting this StopIteration results in a segfault!
@@ -419,8 +463,8 @@ def partition(pred, iterable, tolist=False):
     'Use a predicate to partition entries into false entries and true entries'
     # partition(is_odd, range(10)) --> 0 2 4 6 8   and  1 3 5 7 9
     t1, t2 = itertools.tee(iterable)
-    ifalse = itertools.ifilterfalse(pred, t1)
-    itrue = itertools.ifilter(pred, t2)
+    ifalse = six.moves.filterfalse(pred, t1)
+    itrue = six.moves.filter(pred, t2)
     if tolist:
         return list(ifalse), list(itrue)
     else:
@@ -433,12 +477,12 @@ def merge_config_files(fnames):
     """
     def _load_yaml(fname):
         with open(fname) as in_handle:
-            config = yaml.load(in_handle)
+            config = yaml.safe_load(in_handle)
         return config
     out = _load_yaml(fnames[0])
     for fname in fnames[1:]:
         cur = _load_yaml(fname)
-        for k, v in cur.iteritems():
+        for k, v in cur.items():
             if k in out and isinstance(out[k], dict):
                 out[k].update(v)
             else:
@@ -452,7 +496,7 @@ def deepish_copy(org):
     http://writeonly.wordpress.com/2009/05/07/deepcopy-is-a-pig-for-simple-data/
     """
     out = dict().fromkeys(org)
-    for k, v in org.iteritems():
+    for k, v in org.items():
         if isinstance(v, dict):
             out[k] = deepish_copy(v)
         else:
@@ -464,6 +508,19 @@ def deepish_copy(org):
                 except TypeError:
                     out[k] = v      # ints
     return out
+
+def safe_to_float(x):
+    """Convert to float, handling None and non-float inputs.
+
+    Useful for cleaning complicated output from variant callers.
+    """
+    if x is None:
+        return None
+    else:
+        try:
+            return float(x)
+        except ValueError:
+            return None
 
 def get_in(d, t, default=None):
     """
@@ -483,8 +540,7 @@ def flatten(l):
 
     """
     for el in l:
-        if isinstance(el, collections.Iterable) and not isinstance(el,
-                                                                   basestring):
+        if isinstance(el, (list, tuple)):
             for sub in flatten(el):
                 yield sub
         else:
@@ -499,9 +555,9 @@ def is_sequence(arg):
     example: arg("lol") -> False
 
     """
-    return (not hasattr(arg, "strip") and
-            hasattr(arg, "__getitem__") or
-            hasattr(arg, "__iter__"))
+    return (not is_string(arg) and
+            (hasattr(arg, "__getitem__") or
+             hasattr(arg, "__iter__")))
 
 
 def is_pair(arg):
@@ -512,7 +568,7 @@ def is_pair(arg):
     return is_sequence(arg) and len(arg) == 2
 
 def is_string(arg):
-    return isinstance(arg, basestring)
+    return isinstance(arg, six.string_types)
 
 
 def locate(pattern, root=os.curdir):
@@ -580,8 +636,11 @@ def replace_directory(out_files, dest_dir):
         raise ValueError("in_files must either be a sequence of filenames "
                          "or a string")
 
-def which(program):
+
+def which(program, env=None):
     """ returns the path to an executable or None if it can't be found"""
+    if env is None:
+        env = os.environ.copy()
 
     def is_exe(fpath):
         return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
@@ -591,11 +650,14 @@ def which(program):
         if is_exe(program):
             return program
     else:
-        for path in os.environ["PATH"].split(os.pathsep):
+        for path in env["PATH"].split(os.pathsep):
             exe_file = os.path.join(path, program)
             if is_exe(exe_file):
                 return exe_file
-
+    for path in get_all_conda_bins():
+        exe_file = os.path.join(path, program)
+        if is_exe(exe_file):
+            return exe_file
     return None
 
 def reservoir_sample(stream, num_items, item_parser=lambda x: x):
@@ -632,7 +694,7 @@ def Rscript_cmd():
 
     Prefers Rscript version installed via conda to a system version.
     """
-    rscript = which(os.path.join(os.path.dirname(os.path.realpath(sys.executable)), "Rscript"))
+    rscript = which(os.path.join(get_bcbio_bin(), "Rscript"))
     if rscript:
         return rscript
     else:
@@ -641,9 +703,7 @@ def Rscript_cmd():
 def R_sitelib():
     """Retrieve the R site-library installed with the bcbio installer.
     """
-    from bcbio import install
-    return os.path.join(install.get_defaults().get("tooldir", "/usr/local"),
-                        "lib", "R", "site-library")
+    return os.path.join(os.path.dirname(get_bcbio_bin()), "lib", "R", "library")
 
 def R_package_path(package):
     """
@@ -651,12 +711,12 @@ def R_package_path(package):
     """
     local_sitelib = R_sitelib()
     rscript = Rscript_cmd()
-    cmd = """{rscript} -e '.libPaths(c("{local_sitelib}")); find.package("{package}")'"""
+    cmd = """{rscript} --vanilla -e '.libPaths(c("{local_sitelib}")); find.package("{package}")'"""
     try:
         output = subprocess.check_output(cmd.format(**locals()), shell=True)
     except subprocess.CalledProcessError as e:
         return None
-    for line in output.split("\n"):
+    for line in output.decode().split("\n"):
         if "[1]" not in line:
             continue
         dirname = line.split("[1]")[1].replace("\"", "").strip()
@@ -664,10 +724,56 @@ def R_package_path(package):
             return dirname
     return None
 
+def R_package_resource(package, resource):
+    """
+    return a path to an R package resource, if it is available
+    """
+    package_path = R_package_path(package)
+    if not package_path:
+        return None
+    package_resource = os.path.join(package_path, resource)
+    if not file_exists(package_resource):
+        return None
+    else:
+        return package_resource
+
+def get_java_binpath(cmd=None):
+    """Retrieve path for java to use, handling custom BCBIO_JAVA_HOME
+
+    Defaults to the dirname of cmd, or local anaconda directory
+    """
+    if os.environ.get("BCBIO_JAVA_HOME"):
+        test_cmd = os.path.join(os.environ["BCBIO_JAVA_HOME"], "bin", "java")
+        if os.path.exists(test_cmd):
+            cmd = test_cmd
+    if not cmd:
+        cmd = Rscript_cmd()
+    return os.path.dirname(cmd)
+
+def clear_java_home():
+    """Clear JAVA_HOME environment or reset to BCBIO_JAVA_HOME.
+
+    Avoids accidental java injection but respects custom BCBIO_JAVA_HOME
+    command.
+    """
+    if os.environ.get("BCBIO_JAVA_HOME"):
+        test_cmd = os.path.join(os.environ["BCBIO_JAVA_HOME"], "bin", "java")
+        if os.path.exists(test_cmd):
+            return "export JAVA_HOME=%s" % os.environ["BCBIO_JAVA_HOME"]
+    return "unset JAVA_HOME"
+
+def get_java_clprep(cmd=None):
+    """Correctly prep command line for java commands, setting PATH and unsetting JAVA_HOME.
+    """
+    return "%s && export PATH=%s:\"$PATH\"" % (clear_java_home(), get_java_binpath(cmd))
+
+def get_R_exports():
+    return "unset R_HOME && unset R_LIBS && export PATH=%s:\"$PATH\"" % (os.path.dirname(Rscript_cmd()))
+
 def perl_cmd():
     """Retrieve path to locally installed conda Perl or first in PATH.
     """
-    perl = which(os.path.join(os.path.dirname(os.path.realpath(sys.executable)), "perl"))
+    perl = which(os.path.join(get_bcbio_bin(), "perl"))
     if perl:
         return perl
     else:
@@ -677,17 +783,117 @@ def get_perl_exports(tmpdir=None):
     """Environmental exports to use conda installed perl.
     """
     perl_path = os.path.dirname(perl_cmd())
-    out = "unset PERL5LIB && export PATH=%s:$PATH" % (perl_path)
+    out = "unset PERL5LIB && export PATH=%s:\"$PATH\"" % (perl_path)
     if tmpdir:
         out += " && export TMPDIR=%s" % (tmpdir)
     return out
 
-def local_path_export(at_start=True):
-    path = os.path.dirname(os.path.realpath(sys.executable))
+
+def get_bcbio_env():
+    env = os.environ.copy()
+    env["PATH"] = append_path(get_bcbio_bin(), env['PATH'])
+    return env
+
+
+def append_path(bin, path, at_start=True):
     if at_start:
-        return "export PATH=%s:$PATH && " % (path)
+        tmpl = "{bin}:{path}"
     else:
-        return "export PATH=$PATH:%s && " % (path)
+        tmpl = "{path}:{bin}"
+    return tmpl.format(bin=bin, path=path)
+
+
+def get_bcbio_bin():
+    return os.path.dirname(os.path.realpath(sys.executable))
+
+def get_all_conda_bins():
+    """Retrieve all possible conda bin directories, including environments.
+    """
+    bcbio_bin = get_bcbio_bin()
+    conda_dir = os.path.dirname(bcbio_bin)
+    if os.path.join("anaconda", "envs") in conda_dir:
+        conda_dir = os.path.join(conda_dir[:conda_dir.rfind(os.path.join("anaconda", "envs"))], "anaconda")
+    return [bcbio_bin] + list(glob.glob(os.path.join(conda_dir, "envs", "*", "bin")))
+
+def get_program_python(cmd):
+    """Get the full path to a python version linked to the command.
+
+    Allows finding python based programs in python 2 versus python 3
+    environments.
+    """
+    full_cmd = os.path.realpath(which(cmd))
+    cmd_python = os.path.join(os.path.dirname(full_cmd), "python")
+    env_python = None
+    if "envs" in cmd_python:
+        parts = cmd_python.split(os.sep)
+        env_python = os.path.join(os.sep.join(parts[:parts.index("envs") + 2]), "bin", "python")
+    if os.path.exists(cmd_python):
+        return cmd_python
+    elif env_python and os.path.exists(env_python):
+        return env_python
+    else:
+        return os.path.realpath(sys.executable)
+
+def local_path_export(at_start=True, env_cmd=None):
+    """Retrieve paths to local install, also including environment paths if env_cmd included.
+    """
+    paths = [get_bcbio_bin()]
+    if env_cmd:
+        env_path = os.path.dirname(get_program_python(env_cmd))
+        if env_path not in paths:
+            paths.insert(0, env_path)
+    if at_start:
+        return "export PATH=%s:\"$PATH\" && " % (":".join(paths))
+    else:
+        return "export PATH=\"$PATH\":%s && " % (":".join(paths))
+
+def locale_export():
+    """Exports for dealing with Click-based programs and ASCII/Unicode errors.
+
+    RuntimeError: Click will abort further execution because Python 3 was
+    configured to use ASCII as encoding for the environment.
+    Consult https://click.palletsprojects.com/en/7.x/python3/ for mitigation steps.
+    """
+    locale_to_use = get_locale()
+    return "export LC_ALL=%s && export LANG=%s && " % (locale_to_use, locale_to_use)
+
+def get_locale():
+    """
+    Looks up available locales on the system to find an appropriate one to pick,
+    defaulting to C.UTF-8 which is globally available on newer systems. Prefers
+    C.UTF-8 and en_US encodings, if available
+    """
+    default_locale = "C.UTF-8"
+    preferred_locales = {"c.utf-8", "c.utf8", "en_us.utf-8", "en_us.utf8"}
+    locale_to_use = None
+    try:
+        locales = subprocess.check_output(["locale", "-a"]).decode(errors="ignore").split("\n")
+    except subprocess.CalledProcessError:
+        locales = []
+    # check for preferred locale
+    for locale in locales:
+        if locale.lower() in preferred_locales:
+            locale_to_use = locale
+            break
+    # if preferred locale not available take first UTF-8 locale
+    if not locale_to_use:
+        for locale in locales:
+            if locale.lower().endswith(("utf-8", "utf8")):
+                locale_to_use = locale
+                break
+    # if locale listing not available, try using the default locale
+    if not locale_to_use:
+        locale_to_use = default_locale
+    return locale_to_use
+
+def java_freetype_fix():
+    """Provide workaround for issues FreeType library symbols.
+
+    libfontconfig.so.1: undefined symbol: FT_Done_MM_Var
+
+    Cheap workaround with LD_PRELOAD, I don't know a better one.
+    """
+    return "export LD_PRELOAD=%s/lib/libfreetype.so && " % os.path.dirname(get_bcbio_bin())
 
 def is_gzipped(fname):
     _, ext = os.path.splitext(fname)
@@ -732,11 +938,26 @@ def max_command_length():
     try:
         arg_max = os.sysconf('SC_ARG_MAX')
         env_lines = len(os.environ) * 4
-        env_chars = sum([len(x) + len(y) for x, y in os.environ.iteritems()])
+        env_chars = sum([len(x) + len(y) for x, y in os.environ.items()])
         arg_length = arg_max - env_lines - 2048
     except ValueError:
         arg_length = DEFAULT_MAX_LENGTH
     return arg_length if arg_length > 0 else DEFAULT_MAX_LENGTH
+
+
+def get_abspath(path, pardir=None):
+    if pardir is None:
+        pardir = os.getcwd()
+    path = os.path.expandvars(path)
+    return os.path.normpath(os.path.join(pardir, path))
+
+def sort_filenames(filenames):
+    """
+    sort a list of files by filename only, ignoring the directory names
+    """
+    basenames = [os.path.basename(x) for x in filenames]
+    indexes = [i[0] for i in sorted(enumerate(basenames), key=lambda x:x[1])]
+    return [filenames[x] for x in indexes]
 
 # LazyImport from NIPY
 # https://github.com/nipy/nitime/blob/master/nitime/lazyimports.py
@@ -782,3 +1003,13 @@ class LazyImport(types.ModuleType):
     def __repr__(self):
         return "<module '%s' will be lazily loaded>" %\
                 object.__getattribute__(self,'__name__')
+
+def walk_json(d, func):
+    """ Walk over a parsed JSON nested structure `d`, apply `func` to each leaf element and replace it with result
+    """
+    if isinstance(d, Mapping):
+        return OrderedDict((k, walk_json(v, func)) for k, v in d.items())
+    elif isinstance(d, list):
+        return [walk_json(v, func) for v in d]
+    else:
+        return func(d)
